+++ fastapi_mlflow/_mlflow_types.py
@@ -5,7 +5,7 @@ Copyright (C) 2022, Auto Trader UK
 
 """
 from datetime import date, datetime
-from typing import Union
+from typing import Dict, Optional, Union
 
 from mlflow.types import Schema  # type: ignore
 
@@ -25,20 +25,26 @@ MLFLOW_SIGNATURE_TO_PYTHON_TYPE_MAP = {
 }
 
 
-def build_model_fields(schema):
-    fields = {}
-    if schema.has_input_names():
-        fields.update(
-            {
-                item["name"]: (
-                    MLFLOW_SIGNATURE_TO_PYTHON_TYPE_MAP.get(item["type"]),
-                    ...,  # ... because default value is unknown
-                )
-                for item in schema.to_dict()
-            }
-        )
-    else:
-        rtype = MLFLOW_SIGNATURE_TO_PYTHON_TYPE_MAP.get(schema.numpy_types()[0].name)
-        fields["prediction"] = (rtype, ...)  # ... because default value is unknown
+def get_field(type_name: str, nullable: bool):
+    """
+    :param nullable (bool): Should field be nullable
+    """
+    type_ = MLFLOW_SIGNATURE_TO_PYTHON_TYPE_MAP.get(type_name)
+    if nullable:
+        type_ = Optional[type_]
 
-    return fields
+    field = (type_, ...)  # Ellipsis (...) because default value is unknown
+    return field
+
+
+def build_model_fields(schema, nullable: bool = False) -> Dict:
+    """Return a dict mapping field names -> (type, default).
+
+    :param nullable (bool): Should fields be nullable
+    """
+    if schema.has_input_names():
+        return {
+            item["name"]: get_field(item["type"], nullable) for item in schema.to_dict()
+        }
+    else:
+        return {"prediction": get_field(schema.numpy_types()[0].name, nullable)}